add radiotap debug level and refactor radiotap namespace
