+++ WBLib.cmake
@@ -30,8 +30,8 @@ target_sources(wifibroadcast PRIVATE
         ${CMAKE_CURRENT_LIST_DIR}/src/WBStreamTx.cpp
         ${CMAKE_CURRENT_LIST_DIR}/src/WBTxRx.cpp
         ${CMAKE_CURRENT_LIST_DIR}/src/Ieee80211Header.hpp
-        ${CMAKE_CURRENT_LIST_DIR}/src/RadiotapHeader.hpp
-        ${CMAKE_CURRENT_LIST_DIR}/src/RadiotapHeaderHolder.hpp
+        src/radiotap/RadiotapHeaderTx.hpp
+        src/radiotap/RadiotapHeaderTxHolder.hpp
         ${CMAKE_CURRENT_LIST_DIR}/src/RSSIAccumulator.hpp
         ${CMAKE_CURRENT_LIST_DIR}/src/wifibroadcast_spdlog.cpp
         ${CMAKE_CURRENT_LIST_DIR}/src/Encryption.cpp
