+++ executables/example_pollute.cpp
@@ -44,7 +44,7 @@ int main(int argc, char *const *argv) {
   options_txrx.debug_tx_injection_time= true;
   options_txrx.tx_without_pcap=true;
 
-  auto radiotap_header_holder=std::make_shared<RadiotapHeaderHolder>();
+  auto radiotap_header_holder=std::make_shared<RadiotapHeaderTxHolder>();
   std::shared_ptr<WBTxRx> txrx=std::make_shared<WBTxRx>(cards,options_txrx,radiotap_header_holder);
   // We do not need receive in this mode
   //txrx->start_receiving();
