+++ src/radiotap/README.md
@@ -1,0 +1,2 @@
+Code for dealing with creating radiotap header for injection /
+parsing rx radiotap header.
